Added BottomHeader for new features
Previous Header was split into a base Header component and TopHeader.
